Extract HasBaseName model implementation to mixin interface
