+++ subprojects/platform-c/src/main/java/dev/nokee/platform/c/internal/plugins/CApplicationPlugin.java
@@ -100,6 +100,7 @@ public class CApplicationPlugin implements Plugin<Project> {
 		, ModelBackedHasDevelopmentVariantMixIn<NativeApplication>
 		, ModelBackedTargetMachineAwareComponentMixIn
 		, ModelBackedTargetBuildTypeAwareComponentMixIn
+		, ModelBackedHasBaseNameMixIn
 	{
 		@Override
 		public CSourceSet getCSources() {
